More changes from review:
I didn't like the name of the helper method 'checkForCommit' because it
doesn't indicate that it might commit the event.  I also don't like
'commitEvent' because it might not.  Since JFR events are sort of like a
queue I went with a name from collections and called it 'offer' so using
it is something like 'SocketReadEvent.offer(...)' which seems like it
gets the idea across better.  Also improved the javadoc for it.

Removed the comments about being instrumented by JFR in
Socket.SocketInputStream and Socket.SocketOutputStream.

I went ahead and moved the event commiting out of the finally block so
that we don't emit events when the read/write did not actually happen.
The bugid JDK-8310979 will be used to determine if more should be done
in this area.

The implRead and implWrite were moved up with the other support methods
for read/write.