8315937: Enable parallelism in vmTestbase/nsk/stress/numeric test
'vmTestbase/nsk/stress/numeric' is a small and quick test suite. There seems to be no reason to run these tests exclusively. The tests themselves can be run as performance tests, but they are not executed as such in current configs. We should consider enabling parallelism for them and get improved test performance. Currently it is blocked by 'TEST.properties' with 'exclusiveAccess.dirs' directives in them.

Below are few metrics which shows around 10% improvement in fastdebug mode and around 5% improvement in release mode without any regression:

* fastdebug_before: **72.78s user 20.76s system 272% cpu 34.337 total**
* fastdebug_after : **73.63s user 19.73s system 303% cpu 30.711 total**
* release_before  : **33.42s user 19.42s system 241% cpu 21.898 total**
* release_after   : **33.47s user 18.60s system 255% cpu 20.364 total**