Remove assertions in queue destructor
The queue destructor requires the queue to be empty, but this is
not easy to achieve in cases like ours with a class-level queue.
Not all uses need to have an empty queue at shutdown. This could
also be made optional at instantiation time if we want to keep
the assertions for other users (at the moment only g1 dirty card queue
set).