8068958: Timestamp.from(Instant) should throw when conversion is not possible
Multiplying with `*` never produces `ArithmeticException`, so the catch
in the existing code is never triggered. `Math.multiplyExact` does
produce `ArithmeticException` if the multiplication overflows. So
we can use that, and rethrow `IllegalArgumentException` as the
specification says.

There is a small compatibility risk, in that code may have been relying
on the previous silent overflow, and will now get an exception. But an
exception is surely better than the nonsense results that overflow
produces.

Thanks to Kurt Kluever for the test cases.