The test is for Linux and macOS only
On Windows, the test usually fails with OutOfMemoryException.
Because ShellFolder.invoke() runs on COM thread on Windows,
all the (worker) threads are serialised on COM thread, reducing
the opportunity to concurrently modify fileCache.
On Linux and *headless* macOS, ShellFolder.invoke() runs in the
same thread, therefore all the threads run simultaneously.