JDK-8325264: two compiler/intrinsics/float16 tests fail after JDK-8324724
Fix for two JTREG test failures on OS_X after this commit -
https://github.com/openjdk/jdk/commit/51853f7488afa69c0d14b0e96f1da84822cd83f1.

This is due to no permission to execute the stub code in the code cache.
This patch fixes this problem by acquiring "WXExec" to be able to
execute the instructions in code cache.

Both the tests - compiler/intrinsics/float16/TestAllFloat16ToFloat.java
and compiler/intrinsics/float16/Binary16ConversionNaN.java pass on a
MacOS machine.