8327286
Hi all,

  please review making the flag `NUMAPageScanRate` obsolete.

This option controlled some fairly esoteric performance capability for Solaris; Solaris support has been removed in JDK 15, and the last remaining reference to the option (which was a no-op) in JDK 22.

Testing: local compilation, grepping sources, starting the VM with that option to see that the message shows up.

Thanks,
  Thomas