8327946: containers/docker/TestJFREvents.java fails when host kernel config vm.swappiness=0 after 8325139
According to the docker document(https://docs.docker.com/config/containers/resource_constraints/#--memory-swappiness-details), the default value of --memory-swappiness is inherited from the host machine.
So, when the the kervel config vm.swappiness=0 on the host machine, this testcase will fail, because of docker container can not use swap memory, the deafult value of --memory-swappiness is 0.

Signed-off-by: sendaoYan <yansendao.ysd@alibaba-inc.com>