8328821: Make the ImmutableCollections clear() call consistent
Without overriding clear(), a call to it in an empty map would
just return, as iterator.hasNext() would be false. However if
calling Map.of().clear() throws an exception. To make the
behavior of Map.of().entrySet().clear() consistent, we need to
have an implementation of clear() for the entry set that throws
as well.