Use > 2 condition for error case
There could be two file loader threads active. It depends on
when the currently running thread is interrupted. To avoid false
failures, only snapshots with more than 2 file loader threads
are considered as failure.

For easier failure analysis, the number of snapshots
with 1, 2 and more loader threads is printed to the console.