8330847
Hi all,

  please review this change that fixes survivor rate predictor initialization when expanding the survivor rate prediction tables.

Before JDK-8231579 these predictors were not used when finalizing the collection set at gc start, and were always updated properly at the end of the GC. With that change G1 started using these uninitialized predictors.

There is an assert that should fail in the future if there is a bug.

Testing: tier1-3

Thanks,
  Thomas