Improve handling when the token variant is unknown
Avoid registering CTS algorithms (those depending on CKM_AES_CTS) when
the token CTS variant has not been specified in the configuration. Make
NSS an exception, as we know that it uses the CS1 variant.

Take advantage to extract a pkcs11.Config::parseEnumEntry() method for
a cleaner entry in the main switch statement of pkcs11.Config::parse(),
also slightly improving the error message.

Co-authored-by: Francisco Ferrari <fferrari@redhat.com>
Co-authored-by: Martin Balao <mbalao@redhat.com>