8336912
Hello,

  please review this change to workaround an unintentional division by zero found by ubsan.

The global option `G1ConfidencePercent` allowed zero as confidence for predictions which this change simply disallows: not sure what the response of the VM could be if the user told it that it should have absolutely no confidence in the predictions (not "very low" confidence).

Testing: after this change, the ubsan failure goes away.

Thanks,
  Thomas