fixup! 8341194: [REDO] Implement C2 VectorizedHashCode on AArch64
Eliminate a mismatch between the arrays_hashcode() template and stub.

The template defined v0-v7 & v12-v17 as TEMP  registers but the stub
used v8-v9 which were not on the list. Changing v8-v9 in the stub to
v12-v13 eliminates this mismatch. Other changes in aarch64.ad  is just a
cleanup as some registers on the list are not used by the stub at all
anymore.