8338751: ConfigureNotify behavior has changed in KWin 6.2
Prior to 6.2, kwin used to always send synthetic ConfigureNotify events
regardless whether a real ConfigureNotify event is going to be sent. It
used to do it because of some bugs in GTK2.

In 6.2, that workaround was dropped (it was needed to fix other issues
during interactive resize) so kwin adheres to ICCCM 4.1.5 now.

Unfortunately, it seems like JDK relies on kwin always sending synthetic
configure notify events, which is no longer the case.

This change makes AWT use the TranslateCoordinates request when a real
ConfigureNotify event is received and running kwin as the window manager.
It should work as expected even with kwin < 6.2.