8342782: make listener notifications avoid recursion
This successfully avoids recursion / large stacks, but it negatively impacts performance.

Here is a table showing the relative time it takes to execute one of these methods as the size of the AWTEventMulticaster increases in size:
1     ->  99.5505618%
2     ->  172.972973%
4     ->  289.6749522%
8     ->  205.6451613%
16    ->  191.4498141%
32    ->  132.5908558%
64    ->  147.2392638%
128   ->  117.5475687%
256   ->  118.8359788%
512   ->  102.5245442%
1024  ->  103.8095238%
2048  ->  93.36898396%
4096  ->  95.54210237%
8192  ->  96.96969697%
16384  -> 	96.28610729%
32768  ->  87.19242902%

At 65536+ we see where this change really shines:
The previous code throws a StackOverflowError, and this code doesn't.

However: 65536 nodes is (I assume) an extreme edge case. Slowing down event notifications by 300% for smaller trees is a significant risk that I'm not ready to propose to the OpenJDK project.