8342782: code cleanup for PR
The previous implementation could, technically, traverse 1,000+ nodes, only to eventually return true if level > 500. So now we return true as soon as we detect this condition.

(This is probably only a theoretical gripe, though. Because in reality *when we first hit 500* we'd have already rebalanced, so it'd be very hard to make an unbalanced tree with 1,000 elements...)