8160327: wrap LocaleDateTime.parse in try/catch
I've observed a few cases where the time is not formatted correctly.

Previously: the way the parsing logic was positioned meant the ExifMarkerSegment threw an exception. (Meaning: we lost the ability to parse a thumbnail, all because the timestamp had a formatting error.)

Now: we just leave the timestamp empty if we can't parse it.