8160327: avoid ArrayIndexOutOfBoundsE for unknown data format
I observed one JPEG had a data format not covered by `bytesPerComponent`. I don't know what to do with that information, but let's not throw an ArrayIndexOutOfBoundsException when we get in that situation. We can assume its length is at least 1.

(Unfortunately I can't legally release the JPEG file into public domain as part of a unit test.)