Eliminate no-fault degen penalties
As originally implemented, we apply penalties to the triggering
heuristic every time we experience a degenerated cycle.  This has the
effect of forcing GC triggers to spiral out of control.  This commit
changes the penalty mechanism.  When a degen happens through no fault of
the heuristic triggering mechanism, we do not pile on additional
penalties.  Specifically, we consider that heuristic triggering is not
responsible for a degenerated cycle that is associated with a GC that
began immediately following the end of the previous GC cycle.