8348659: AArch64: IR rule failure with compiler/loopopts/superword/TestSplitPacks.java
"test5a" in this file fails on Graviton3 (32B, SVE) as the compiler fails
to match IR rules for vector size 2. This is because the
minimum vector size for aarch64 machines is 8B and it does not support
generation of vectors of 2 short values.

Modified the IR rules to have two separate rules - one for sse4.1 and
another for sve.

The test now passes on Graviton3.