Respond to reviewer feedback
In testing suggested refinements, I discovered a bug in original
implementation.  ShenandoahFreeSet::capacity() does not represent the
size of young generation.  It represents the total size of the young
regions that had available memory at the time we most recently rebuilt
the ShenandoahFreeSet.

I am rerunning the performance tests following this suggested change.