8349943: [JMH] Use jvmArgs consistently
As is suggested in JDK-8342958, `jvmArgs` should be used consistently in
microbenchmarks to 'align with the intuition that when you use
jvmArgsAppend/-Prepend intent is to add to a set of existing flags,
while if you supply jvmArgs intent is "run with these and nothing else"'.

All the previous flags were aligned in
https://github.com/openjdk/jdk/pull/21683, while some recent tests use
inconsistent `jvmArgs` again. We update them to keep the consistency.

Change-Id: Ia52af6d2132c561f58bf07bc2f78eed41134d2d8