fixup: don't modify the value in vsrc
Fix reduce_mul_integral_gt128b() so it doesn't modify vsrc. With this
change, the result of recursive folding is held in vtmp1. To be able to
pass this intermediate result to reduce_mul_integral_le128b(), we would
have to use another temporary FloatRegister, as vtmp1 would essentially
act as vsrc. It's possible to get around this however:
reduce_mul_integral_le128b() is modified so it's possible to pass
matching vsrc and vtmp2 arguments. By doing this, we save ourselves a
temporary register in rules that match to reduce_mul_integral_gt128b().