8349944: [JMH] sun.misc.UnsafeOps cannot access class jdk.internal.misc.Unsafe
The UnsafeOps JMH benchmark fails with the following error:

      ```
      java.lang.IllegalAccessError: class org.openjdk.bench.sun.misc.UnsafeOps (in unnamed module @0x520a3426) cannot access class jdk.internal.misc.Unsafe (in module java.base) because module java.base does not export jdk.internal.misc to unnamed module @0x520a3426
      ```

Since this micro-benchmark is created for `sun.misc.Unsafe` rather than
`jdk.internal.misc.Unsafe`, we change it back as before JDK-8332744.

Note that even it will raise "proprietary API" warnings after this
patch, it is acceptable because the relevant APIs are bound for removal
for the integrity of the platform.

Change-Id: Ia7c57c2ca09af4b2b3c6cc10ef4ae5a9f3c38a4c