8160327: replacing the "sony-d700" image
The origins of that image were unknown, so we weren't sure if we had the rights to store it in the OpenJDK repo.

I couldn't figure out how to create this kind of uncompressed thumbnail from an image editing app, so I spliced this new file together manually in a hex editor using the sony-d700 image as a blueprint.