Improving NPE message, to ease debugging
Went down an unnecessary rabbit hole at work today. A better error message would have made this easier to track down. It is my fault, but this is a zero-cost change -- it's functionally the same as what it was before.

I know I am not following the proper processes, but hopefully we are good with just letting this one through? It's literally just adding a better message to a NPE and breaking out the cases.