8351322: Parameterize link option for pthreads
Replace hardcoded instances of `-lpthread` with `$(LIBPTHREAD)`, so that
it's possible to parameterize this for platforms that use different
flags for enabling posix threads.

This work is a continuation of the work done by Greg Lewis in [1], but
generalized for the full JDK, and set at the configure stage.

Sponsored by: The FreeBSD Foundation
Co-authored-by: Greg Lewis <glewis@eyesbeyond.com>

[1]: https://github.com/battleblow/jdk23u/commit/dbd90aa8ab0b7f5e4865864a7c63d975daacabf4