8351110: avoid writing invalid JPEG segments
IMO the most serious part of this ticket is that the caller receives no indication that the file that was written is corrupt.

With this change the caller will at least see an IIOException.

This does not satisfactorily resolve the ticket (we want some sort of non-corrupt JPEG), but it's a start.

This placement may be invasive, but I think it's appropriate. I'm having a hard time imagining a scenario where it's safe/ok to call `write2bytes` for -1 or 70,000.