8351110: write thumbnail using JFIFThumbJPEG
This is my preferred resolution to this ticket, however I'm worried it's too unorthodox so I'll revert it shortly.

Now if the thumbnail is going to be too large: we switch from writing an RGB-style thumbnail to a JPEG-compressed thumbnail.

In this ticket's case: now the call to `ImageIO.write(..)` "just works": the caller gets a valid JPEG file with the entire thumbnail.