8351110: write thumbnail using reduced thumbnail size
We were already clipping the thumbnail width/height to 255px.

Now we're also clipping the thumbnail width/height if the total number of bytes we write will exceed the segment capacity.

We may (?) want to introduce a separate warning constant (other than WARNING_THUMB_CLIPPED?), which may involve a CSR.